backend/requirements.txt
Django
djangorestframework
django-cors-headers
PyJWT
bcrypt
python-decouple
mysqlclient
psycopg2-binary
django-ratelimit
Pillow
django-extensions